// src/shared/actionNames.enum.ts
export enum ActionName {
  CREATE = 'Create',
  RENAME = 'Rename',
  SPLIT = 'Split',
}

export enum ActionType {
  CREATE_DOCUMENT = 'Create document',
  CREATE_PAGE = 'Create page',
  DELETE_DOCUMENT = 'Delete document',
  DELETE_PAGE = 'Delete page',
  RENAME = 'Rename',
  UPDATE_PROPERTIES = 'Update properties'
}

export enum SortType {
  NONE = 'None',
  ASC = 'Asc',
  DESC = 'Desc'
}